Changes to do (1) auto-calibration with chosen values of a, b, T, and (2) Changes to do the edge re-mapping properly for doing edge speed calculations. Note that there is functionality that is commented out but is now in the code.

+++ LivingCity/traffic/b18EdgeData.h
@@ -34,7 +34,7 @@ struct B18EdgeData {
   float length;
   float maxSpeedMperSec;
   uint nextIntersMapped;
-  int curr_iter_num_cars = 0;
+  float curr_iter_num_cars = 0;
   float curr_iter_cum_vel = 0;
   float curr_cum_vel = 0;
 };
